// happytrip_frontend/src/app/models/getAllBusesByRouteDateModel.ts
export interface getAllBusesResponse {
    busId: number
    busName: string
    source: string
    destination: string
    boardingPoint: string
    droppingPoint: string
    departureDate: string
    departureTime: string
    amenities: string
    coachType: string
    totalAvailableSeats: string
    classAvailable: string
    bookingPolicies: string
    classWithFare: string
  }